// src/app/model/employee.ts
export class Employee {
    id: string;
    code: string;
    name: string;
    password: string;
    birthday: Date;
    currentaddress:string;
    nativeaddress: string;
    CMND: string;
    phone: string;
    email: string;
    isMarried: boolean;
    haveChildren: boolean;
    postion: string;
}